requirements.txt
crewai
crewai-tools
langchain_groq
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
slack_sdk
flask